+++ Plugins/OpenApi/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/openapi/spec/OpenApiGenerator.java
@@ -67,7 +67,7 @@ public class OpenApiGenerator {
         paths.put(context.getBase(), basePath);
 
         for (EntityType entityType : modelRegistry.getEntityTypes()) {
-            addPathsForSet(document, 0, paths, context.getBase(), entityType, context);
+            addPathsForSet(document, 0, paths, context.getBase(), entityType, entityType.plural, context);
         }
         return document;
     }
@@ -407,19 +407,19 @@ public class OpenApiGenerator {
         }
     }
 
-    private static void addPathsForSet(OADoc document, int level, Map<String, OAPath> paths, String base, EntityType entityType, GeneratorContext options) {
-        String path = base + "/" + entityType.plural;
+    private static void addPathsForSet(OADoc document, int level, Map<String, OAPath> paths, String base, EntityType entityType, String setName, GeneratorContext options) {
+        String path = base + "/" + setName;
         OAPath pathCollection = createPathForSet(options, path, entityType, level > 0);
         paths.put(path, pathCollection);
 
         if (options.isAddRef()) {
-            String refPath = base + "/" + entityType.plural + "/$ref";
+            String refPath = path + "/$ref";
             OAPath pathRef = createPathForSetRef(options, refPath, entityType, level > 0);
             paths.put(refPath, pathRef);
         }
 
         if (level < options.getRecurse()) {
-            String baseId = base + "/" + entityType.plural + "({entityId})";
+            String baseId = path + "({entityId})";
             OAPath pathBaseId = createPathForEntity(options, baseId, entityType);
             paths.put(baseId, pathBaseId);
             addPathsForEntity(document, level, paths, baseId, entityType, options);
@@ -432,13 +432,14 @@ public class OpenApiGenerator {
         }
         for (NavigationProperty navProp : entityType.getNavigationSets()) {
             if (level < options.getRecurse()) {
-                addPathsForSet(document, level + 1, paths, base, navProp.getEntityType(), options);
+                final EntityType targetType = navProp.getEntityType();
+                addPathsForSet(document, level + 1, paths, base, targetType, navProp.getName(), options);
             }
         }
         for (NavigationProperty navProp : entityType.getNavigationEntities()) {
             if (level < options.getRecurse()) {
                 EntityType type = navProp.getEntityType();
-                String baseName = base + "/" + type.entityName;
+                String baseName = base + "/" + navProp.getName();
                 OAPath paPath = createPathForEntity(options, baseName, type);
                 paths.put(baseName, paPath);
                 addPathsForEntity(document, level + 1, paths, baseName, type, options);